+++ src/modules/methods/ni-mcit-d-01.service.ts
@@ -44,6 +44,7 @@ import {
 import { countDecimals } from 'src/utils/countDecimal'
 import { MethodsService } from './methods.service'
 import { formatCertCode, formatQuoteCode } from 'src/utils/generateCertCode'
+import { EnginesService } from '../engines/engines.service'
 
 @Injectable()
 export class NI_MCIT_D_01Service {
@@ -84,6 +85,9 @@ export class NI_MCIT_D_01Service {
 
     @Inject(forwardRef(() => MethodsService))
     private readonly methodService: MethodsService,
+
+    @Inject(forwardRef(() => EnginesService))
+    private readonly enginesService: EnginesService,
   ) {}
 
   async create() {
@@ -576,11 +580,16 @@ export class NI_MCIT_D_01Service {
     //   )
     // }
 
-    const filePath = path.join(
-      __dirname,
-      `../mail/templates/excels/ni_mcit_d_01.xlsx`,
-    )
+    const enginePath =
+      await this.enginesService.getPathByCalibrationMethodAndPattern(
+        'NI-MCIT-D-01',
+      )
 
+    const filePath = path.join(__dirname, enginePath)
+
+    if (!enginePath) {
+      return handleInternalServerError('No se encontró la ruta del motor')
+    }
     try {
       if (fs.existsSync(method.certificate_url)) {
         fs.unlinkSync(method.certificate_url)
