+++ src/modules/methods/ni-mcit-t-01.service.ts
@@ -31,6 +31,7 @@ import {
   formatSameNumberCertification,
 } from 'src/utils/formatNumberCertification'
 import { countDecimals } from 'src/utils/countDecimal'
+import { EnginesService } from '../engines/engines.service'
 
 @Injectable()
 export class NI_MCIT_T_01Service {
@@ -63,6 +64,9 @@ export class NI_MCIT_T_01Service {
 
     @Inject(forwardRef(() => MethodsService))
     private readonly methodService: MethodsService,
+
+    @Inject(forwardRef(() => EnginesService))
+    private readonly enginesService: EnginesService,
   ) {}
 
   async create() {
@@ -375,10 +379,16 @@ export class NI_MCIT_T_01Service {
     }
 
     try {
-      const filePath = path.join(
-        __dirname,
-        '../mail/templates/excels/ni_mcit_t_01.xlsx',
-      )
+      const enginePath =
+        await this.enginesService.getPathByCalibrationMethodAndPattern(
+          'NI-MCIT-T-01',
+        )
+
+      const filePath = path.join(__dirname, enginePath)
+
+      if (!enginePath) {
+        return handleInternalServerError('No se encontró la ruta del motor')
+      }
 
       if (fs.existsSync(method.certificate_url)) {
         fs.unlinkSync(method.certificate_url)
