+++ src/modules/methods/ni-mcit-t-05.service.ts
@@ -31,6 +31,7 @@ import {
   formatSameNumberCertification,
 } from 'src/utils/formatNumberCertification'
 import { countDecimals } from 'src/utils/countDecimal'
+import { EnginesService } from '../engines/engines.service'
 
 @Injectable()
 export class NI_MCIT_T_05Service {
@@ -63,6 +64,9 @@ export class NI_MCIT_T_05Service {
     private readonly dataSource: DataSource,
     @Inject(forwardRef(() => ActivitiesService))
     private activitiesService: ActivitiesService,
+
+    @Inject(forwardRef(() => EnginesService))
+    private readonly enginesService: EnginesService,
   ) {}
 
   async create() {
@@ -434,11 +438,16 @@ export class NI_MCIT_T_05Service {
     }
 
     try {
-      const filePath = path.join(
-        __dirname,
-        '../mail/templates/excels/ni_mcit_t_05.xlsx',
-      )
+      const enginePath =
+        await this.enginesService.getPathByCalibrationMethodAndPattern(
+          'NI-MCIT-T-05',
+        )
 
+      const filePath = path.join(__dirname, enginePath)
+
+      if (!enginePath) {
+        return handleInternalServerError('No se encontró la ruta del motor')
+      }
       if (fs.existsSync(method.certificate_url)) {
         fs.unlinkSync(method.certificate_url)
       }
