+++ src/modules/methods/ni-mcit-v-01.service.ts
@@ -28,6 +28,7 @@ import { CertificationDetailsDto } from './dto/NI_MCIT_P_01/certification_detail
 import { formatCertCode, formatQuoteCode } from 'src/utils/generateCertCode'
 import { formatNumberCertification } from 'src/utils/formatNumberCertification'
 import { countDecimals } from 'src/utils/countDecimal'
+import { EnginesService } from '../engines/engines.service'
 
 @Injectable()
 export class NI_MCIT_V_01Service {
@@ -60,6 +61,9 @@ export class NI_MCIT_V_01Service {
     private readonly dataSource: DataSource,
     @Inject(forwardRef(() => ActivitiesService))
     private activitiesService: ActivitiesService,
+
+    @Inject(forwardRef(() => EnginesService))
+    private readonly enginesService: EnginesService,
   ) {}
 
   async create() {
@@ -434,10 +438,16 @@ export class NI_MCIT_V_01Service {
     }
 
     try {
-      const filePath = path.join(
-        __dirname,
-        '../mail/templates/excels/ni_mcit_v_01.xlsx',
-      )
+      const enginePath =
+        await this.enginesService.getPathByCalibrationMethodAndPattern(
+          'NI-MCIT-V-01',
+        )
+
+      const filePath = path.join(__dirname, enginePath)
+
+      if (!enginePath) {
+        return handleInternalServerError('No se encontró la ruta del motor')
+      }
 
       if (fs.existsSync(method.certificate_url)) {
         fs.unlinkSync(method.certificate_url)
